crawler/requirements.txt
beautifulsoup4>=4.12.0
boto3>=1.34.0
curl_cffi>=0.6.0
h2>=4.1.0
httpx>=0.26.0
lxml>=4.9.0
pillow>=10.0.0
playwright>=1.40.0
pydantic>=2.0.0
anticaptchaofficial>=1.0.40
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.0.0
